+++ Game1.cs
@@ -9,6 +9,7 @@ using Project.Tiles;
 using System.Collections.Generic;
 using Project.Sprites.Characters.Enemy;
 using Project.Sprites;
+using Project.Scenes;
 
 namespace Project
 {
@@ -17,23 +18,7 @@ namespace Project
         private GraphicsDeviceManager graphics;
         private SpriteBatch spriteBatch;
 
-        Camera camera = new Camera();
-
-        private Texture2D playerTexture;
-        private Texture2D wizardTexture;
-        private Texture2D knightTexture;
-        private Texture2D fairyTexture;
-        private Texture2D healthTexture;
-        static public Texture2D swordTexture;
-        static public Texture2D fireballTexture;
-        static public Texture2D tilesTexture;
-
-        static public int tileSize = 16;
-
-        public static Player player;
-        private PlayerSword playerSword;
-        public static List<Enemy> enemies;
-        public static MapManager mapManager = new MapManager();
+        public SceneManager SceneManager { get; set; }
 
         public Game1()
         {
@@ -44,61 +29,33 @@ namespace Project
 
         protected override void Initialize()
         {
+            SceneManager = new SceneManager();
+
+            SceneManager.AddScene(SceneType.MainMenu, new StartScreen(this));
+            SceneManager.AddScene(SceneType.Game, new GameScene(this));
+
+            SceneManager.SetScene(SceneType.MainMenu);
+
             base.Initialize();
 
             ScreenManager.Setup(graphics, Window);
-
-            player = new Player(playerTexture, new KeyboardReader(), healthTexture, camera);
-            playerSword = new PlayerSword(swordTexture, player);
-            enemies = new List<Enemy>();
-            enemies.Add(new Wizard(wizardTexture, healthTexture));
-            enemies.Add(new Knight(knightTexture, healthTexture));
-            enemies.Add(new Fairy(fairyTexture, healthTexture));
         }
 
         protected override void LoadContent()
         {
             spriteBatch = new SpriteBatch(GraphicsDevice);
-
-            playerTexture = Content.Load<Texture2D>("lancelot_");
-            swordTexture = Content.Load<Texture2D>("excalibur_");
-            fireballTexture = Content.Load<Texture2D>("FB001");
-            wizardTexture = Content.Load<Texture2D>("merlin_"); 
-            knightTexture = Content.Load<Texture2D>("mordred_");
-            fairyTexture = Content.Load<Texture2D>("morganLeFay_");
-            healthTexture = Content.Load<Texture2D>("heart");
-            tilesTexture = Content.Load<Texture2D>("Dungeon tileset");
-
-            mapManager.LoadMap("Content/Maps/Map-1.txt");
         }
 
         protected override void Update(GameTime gameTime)
         {
             if (GamePad.GetState(PlayerIndex.One).Buttons.Back == ButtonState.Pressed || Keyboard.GetState().IsKeyDown(Keys.Escape))
-                Exit();
+                SceneManager.SetScene(SceneType.MainMenu);
 
-            camera.Follow(player);
-            player.Update(gameTime);
-            playerSword.Update(gameTime);
+            if (GamePad.GetState(PlayerIndex.One).Buttons.Back == ButtonState.Pressed || Keyboard.GetState().IsKeyDown(Keys.Escape))
+                if(SceneManager.CurrentScene == SceneManager.Scenes[SceneType.MainMenu])
+                    Exit();
 
-            foreach (Enemy enemy in enemies)
-            {
-                switch(enemy)
-                {
-                    case Wizard wizard:
-                        wizard.Update(gameTime);
-                        break;
-                    case Knight knight:
-                        knight.Update(gameTime);
-                        break;
-                    case Fairy fairy:
-                        fairy.Update(gameTime);
-                        break;
-                    default:
-                        enemy.Update(gameTime);
-                        break;
-                }
-            }
+            SceneManager.Update(gameTime);
 
             base.Update(gameTime);
         }
@@ -107,37 +64,8 @@ namespace Project
         {
             GraphicsDevice.Clear(Color.CornflowerBlue);
 
-            spriteBatch.Begin(
-                transformMatrix: camera.Transform, 
-                sortMode: SpriteSortMode.FrontToBack,
-                samplerState: SamplerState.PointClamp
-            );
-            mapManager.Draw(spriteBatch);
-                        
-            player.Draw(spriteBatch);
-            playerSword.Draw(spriteBatch);
-
-            foreach (Enemy enemy in enemies)
-            {
-                switch (enemy)
-                {
-                    case Wizard wizard:
-                        wizard.Draw(spriteBatch);
-                        break;
-                    case Knight knight:
-                        knight.Draw(spriteBatch);
-                        break;
-                    case Fairy fairy:
-                        fairy.Draw(spriteBatch);
-                        break;
-                    default:
-                        enemy.Draw(spriteBatch);
-                        break;
-                }
-            }
-
-            spriteBatch.End();
-
+            SceneManager.Draw(spriteBatch);
+            
             base.Draw(gameTime);
         }
     }