+++ Sprites/Characters/Enemy/Fairy.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 using Project.Managers;
+using Project.Scenes;
 using Project.UI;
 using System;
 
@@ -10,7 +11,7 @@ namespace Project.Sprites.Characters.Enemy
     {
         public Fairy(Texture2D texture, Texture2D heartTexture) : base(texture, heartTexture)
         {
-            Position = Game1.mapManager.GetCurrentMap().MidOfMap;
+            Position = GameScene.mapManager.GetCurrentMap().MidOfMap;
             Speed = new Vector2(1, 1);
             Health = new Health(1, heartTexture);
             Damage = 0;
@@ -23,7 +24,7 @@ namespace Project.Sprites.Characters.Enemy
 
         internal override void Move()
         {
-            Vector2 direction = Position - Game1.player.Position;
+            Vector2 direction = Position - GameScene.player.Position;
 
             if (direction != Vector2.Zero)
             {